+++ doc/changelog.md
@@ -1,5 +1,17 @@
 # Changelog
 
+### 0.1.8
+
+- Added `scipy` as a required module
+
+### `Transport`
+
+- There are always 8 target objects in the scene (previously, there are 8-12 target objects)
+- Target objects can spawn in any room (previously, they spawned in the same room)
+- Fixed: Target objects sometimes spawn on the edges of the occupancy map
+- Fixed: Containers sometimes spawn in very small rooms
+- Added optional parameter `fov` to the constructor
+
 ## 0.1.7
 
 ### `Transport`